jamai-llm/requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bitarray==3.0.0
blinker==1.8.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
document==1.0
docx==0.2.4
exceptiongroup==1.2.2
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jamaibase==0.2.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
loguru==0.7.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.10.0
numpy==1.26.4
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pdfreader==0.1.15
pillow==10.4.0
protobuf==5.28.3
pyarrow==17.0.0
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
srsly==2.4.8
streamlit==1.39.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
win32-setctime==1.1.0